backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-ollama
chromadb
python-dotenv
rank_bm25
sentence-transformers
